+++ src/com.radiant.engine/bundle/Systems/UI/UIWindow/Inspector.cs
@@ -577,8 +577,8 @@ public class Inspector : core.System
         ComputeAllLayouts();
 
         var Scale = Matrix.CreateScale(
-            (float)Renderer.ScreenWidth / Renderer.VirtualWidth,
-            (float)Renderer.ScreenHeight / Renderer.VirtualHeight,
+            Renderer.ScreenWidth / Renderer.VirtualWidth,
+            Renderer.ScreenHeight / Renderer.VirtualHeight,
             1f);
 
         var CurrentMouse = Mouse.GetState();
